Intro security checks + PR pipeline (#4)
Previously, the solution structure was fixed and the code was revamped.
This pull request includes updates to the build files, specifically the
base image build file and the general build file. Additionally, a new
build file was added. The changes ensure that the build process is
accurate and up-to-date.

Add PR pipeline and main branch release pipelines

---------